Update apply_transformations.py
the transformations are applied to each patient's CT scan and the labels. The transformed patient CT scans are saved in a directory named after the patient within the Results folder. The transformed labels for each patient are saved in a subdirectory named Transformed_Labels within each patient's directory in Results.